+++ src/main/java/lotto/Application.java
@@ -1,23 +1,20 @@
 package lotto;
 
-import java.util.ArrayList;
+import lotto.view.InputView;
+import lotto.view.OutputView;
+
 import java.util.HashSet;
-import java.util.InputMismatchException;
 import java.util.List;
-import java.util.Scanner;
 
 public class Application {
     public static void main(String[] args) {
         List<Lotto> lottos = purchaseLottos();
-        System.out.println("%d개를 구매했습니다.".formatted(lottos.size()));
-        for (Lotto lotto : lottos) {
-            System.out.println(lotto.getNumbers());
-        }
+        OutputView.printLottos(lottos);
         WinningNumbers winningNumbers = getWinningNumbers();
         int bonusNumber = getBonusNumber(winningNumbers);
         WinningStatistics winningStatistics = LottoManager.calculateStatistics(lottos, winningNumbers, bonusNumber);
         double returnRate = winningStatistics.calculateReturnRate(lottos.size() * LottoManager.LOTTO_UNIT_PRICE);
-        printWinningStatistics(winningStatistics, returnRate);
+        OutputView.printWinningStatistics(winningStatistics, returnRate);
     }
 
     private static List<Lotto> purchaseLottos() {
@@ -25,36 +22,27 @@ public class Application {
             int purchaseAmount = getPurchaseAmount();
             return LottoManager.purchase(purchaseAmount);
         } catch (IllegalArgumentException e) {
-            System.out.println(e.getMessage());
+            OutputView.printErrorMessage(e.getMessage());
             return purchaseLottos();
         }
     }
 
     private static int getPurchaseAmount() {
         try {
-            System.out.println("구입금액을 입력해 주세요.");
-            Scanner scanner = new Scanner(System.in);
-            return scanner.nextInt();
-        } catch (InputMismatchException e) {
-            System.out.println("구입금액은 숫자여야 합니다.");
+            return InputView.inputPurchaseAmount();
+        } catch (NumberFormatException e) {
+            OutputView.printErrorMessage("구입금액은 숫자여야 합니다.");
             return getPurchaseAmount();
         }
     }
 
     private static WinningNumbers getWinningNumbers() {
         try {
-            System.out.println("\n지난 주 당첨 번호를 입력해 주세요.");
-            Scanner scanner = new Scanner(System.in);
-            String input = scanner.nextLine();
-            String[] inputWinningNumbers = input.split(",");
-            List<Integer> winningNumbers = new ArrayList<>();
-            for (String inputWinningNumber : inputWinningNumbers) {
-                winningNumbers.add(Integer.parseInt(inputWinningNumber.trim()));
-            }
+            List<Integer> winningNumbers = InputView.inputWinningNumbers();
             validateWinningNumbers(winningNumbers);
             return new WinningNumbers(winningNumbers);
         } catch (IllegalArgumentException e) {
-            System.out.println(e.getMessage());
+            OutputView.printErrorMessage(e.getMessage());
             return getWinningNumbers();
         }
     }
@@ -71,14 +59,12 @@ public class Application {
 
     private static int getBonusNumber(final WinningNumbers winningNumbers) {
         try {
-            System.out.println("보너스 볼을 입력해 주세요.");
-            Scanner scanner = new Scanner(System.in);
-            int bonusNumber = scanner.nextInt();
+            int bonusNumber = InputView.inputBonusNumber();
             validateLottoNumber(bonusNumber);
             winningNumbers.validateBonusNumberDuplicated(bonusNumber);
             return bonusNumber;
         } catch (IllegalArgumentException e) {
-            System.out.println(e.getMessage());
+            OutputView.printErrorMessage(e.getMessage());
             return getBonusNumber(winningNumbers);
         }
     }
@@ -87,38 +73,5 @@ public class Application {
         if (lottoNumber < 1 || lottoNumber > 45) {
             throw new IllegalArgumentException("로또 번호는 1 ~ 45 사이여야 합니다.");
         }
-    }
-
-    private static void printWinningStatistics(WinningStatistics winningStatistics, double returnRate) {
-        System.out.println("\n당첨 통계");
-        System.out.println("---------");
-
-        for (Prize prize : Prize.values()) {
-            printDetail(prize, winningStatistics.getPrizeCount(prize));
-        }
-
-        printReturnRate(returnRate);
-    }
-
-    private static void printReturnRate(double returnRate) {
-        System.out.print(String.format("총 수익률은 %.2f입니다.", returnRate));
-        if (returnRate < 1) {
-            System.out.println("(기준이 1이기 때문에 결과적으로 손해라는 의미임)");
-        }
-    }
-
-    private static void printDetail(Prize prize, int getPrizeCount) {
-        if (prize == Prize.NONE) {
-            return;
-        }
-
-        if (prize == Prize.SECOND) {
-            System.out.println(String.format("%d개 일치, 보너스 볼 일치(%d원) - %d개",
-                    prize.getMatchCount(), prize.getWinningAmount(), getPrizeCount));
-            return;
-        }
-
-        System.out.println(String.format("%d개 일치 (%d원) - %d개",
-                prize.getMatchCount(), prize.getWinningAmount(), getPrizeCount));
     }
 }
